src/requirements.txt
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-functions==1.18.0